Make column reader check reference count before writing (#919)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/919

Column reader (not the Selective one) didn't check if anyone else holds reference to the vector or buffer. That has resulted in cases that vector referenced by client been silently updated by the reader causing corruption.

Update the reader to check that. The bulk of the change is very mechanical:
1. The first thing to call now is resetIfWrongType(), it checks if vector is single referenced.
2. One that is done, it checks nulls, followed by other buffers
3. When passing VectorPtr to child readers in complex types, the code is changed to make the vectors are single referenced, so child reader doesn't reallocate if so.

Reviewed By: Magoja, HuamengJiang

Differential Revision: D33756784

fbshipit-source-id: 0e726bbcf1cc28c79bed3245165695b2809e4eef