Test peeling with LazyVectors that produce dictionaries
A LazyVector produces a dictionary for a string dictionary column and
for a sparsely accessed column. In the sparse column case, the values
are contiguous but the dictionary maps them to non-contiguous places
as expected by the call site.

When a LazyVector is an argument to a function, it gets loaded,
becomes a dictionary and is then peeled. After this, the function can
return a constant, wich is then re-wrapped in the dictionary. This
gets peeled again by the next outer function, which should run on a
single value since the dictionary is over a constant. We fix the
detection of all constant arguments in applyFunctionWithPeeling. An
argument can start as a wrapper and become a constant, which
possibility was not considered, leading to failing the check for
single active row in Expr::applySingleConstArgVectorFunction.