Add unit test for grouped/bucketed execution. (#1011)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1011

Adding the unist test to test out-of-sync addition of splits and split groups, also for partial split groups and for adding splits before the Task start.
Added a new Task state (kNone) to reflect that we haven't started yet.
Updated exec task utils classes to be able to work with split groups.

Differential Revision: D34088826

fbshipit-source-id: cc70d6bdb68beb5a201cbc261d45a2e8ccb6a07b