Fix LazyVector in dictionary
LazyVectors can b wrapped in dictionaries by joins and filters. Loading the LazyVector for the appropriate rows cannot be done by RowVector::loadedChild because the lazy vector is not the immediate child.

Fix hash join to load LazyVectors for the whole input when these are
operands of a join filter and the filter may be invoked multiple
times. Generally, inputs of joins where the input may be referenced in
more than one batch of output need to be loaded by the join. Otherwise
a downstream Operator could load the LazyVector for a subset of rows
and in the next batch would be missing values for other rows. Load
only for rows that pass the join or all rows for left/full join.

Fix EvalCtx::ensureFieldLoaded to collapse dictionaries inside
dictionaries into a single level of dictionary. Note that if we check
dictionary indices when building dictionary wrappers for selected
rows, inner dictionaries may reference base data rows that outer
dictionaries do not reference. Therefore, when loading a LazyVector
wrapped inside multiple dictionaries, we translate selected rows of
the outermost dictionary into distinct rows in the base data and then
load the base data for these rows. We then wrap the base data in one
level of dictionary.

This operation applies to loading data for expressions as well as
input vectors that Operators access. This does not apply to columns
operators pass through. Operators like PartitionedOutput that process
their whole input in one pass can use DecodedVector for loading
LazyVectors and do not need extra attention to this.