Return ConstantVector column values with the right size
ConstantVector has a size that is set to what is appropriate for the
use context. The constant itself is dimensionless but the vector that
represents an instance of it has a dimension so as to be consistent
with other vectors.

When a column with a constant encoded value is projected out, the
constant is not wrapped in a dictionary but is just copied to a
constant of a different size. Then, when the constant is combined with
a dictionary wrapped column in an expression, the values of the
dictionary and the constant have a different size. This leads to
incorrect results and access of uninitialized, i.e. crash for strings,
when the constant is made writable (flattened) for the wrong number of
elements.

select count(if (c > x, ds, null)) from table where z + 1 > 2;

ds is a partitioning column, hence represented by a constant
vector. The other columns are flat. After the filter, c and x are
wrapped in a dictionary. ds is an unwrapped constant with the size
matching the count of passed rows from z + 1 > 2. The condition in if
is true for some and false for others. We get a constant vector of ds
which is then flattened and set to null for the rows where c > x was
false. This is done in the peeled domain, where vectors are at the
size they had before the filter. But ds is at the size after the
filter and we have therefore uninitialized positions at the end of
flattened ds, unless we return ds at the cardinality in the peeled
(pre-filter) domain.

To effect this change, we make FieldReference resize ConstantVectors
it returns. We do the resize in place for singly referenced and make a
copy otherwise. In order to distinguish singly referenced from
multiply referenced, we make EvalCtx::getField return const VectorPtr&
instead of VectorPtr. Like this the return does not affect reference
count and is also a little faster.

We fix assignment of all-null data to string vectors. The all-null
source vector could be of unknown type and hence would not have string
buffers.