Add Json Presto type and custom cast logic for bigint to json (#940)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/940

This is the first diff of supporting `CAST(... AS JSON)` Presto function in Velox. This diff registers a Json custom type with a custom cast logic and extends `CastExpr` to use this cast logic for Json-related cast operations. The custom cast logic in this diff only supports casts from bigint to json and I'll add the support for other types in a separate diff later.

Differential Revision: https://www.internalfb.com/diff/D33826288?entry_point=27

fbshipit-source-id: 4a1f61df24e8227de51e61b4cce9f85e14ee6f88