Enable memory tracking by default
Default QueryCtx was using memory memory with memory tracking disabled. This
change enables memory tracking and adds a quick santiy check to HashJoinTest.

I noticed that FilterProject operator doesn't report memory usage correctly
because it uses ExecCtx which allocates memory from the Driver's, not
Operator's memory pool. This needs to be fixed in a follow-up PR.