Support 'accuracy' argument in approx_percentile() presto aggregate. (#1046)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1046

Support 'accuracy' argument in approx_percentile() presto aggregate.
'Accuracy' is used to control ... the accuracy of the approximate aggregate.
The less - the more accurate the result (and more expensive to run it).
You can find description here (https://prestodb.io/docs/current/functions/aggregate.html) or in the rst changes in this diff.
We cannot do the same as Presto does, as our aggregation based on TDigest and Presto uses QDigest.
But we do similar thing - we control TDigest's maxSize by the 'accuracy' argument.
By default it is 100 (like TDigest) has it.

Differential Revision: D34234874

fbshipit-source-id: 32912d9ccafbe6f415d4385857d52dd27eeb95f2