Add benchmark for Variadic Args (#1003)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1003

Adding a benchmark to compare a Vector function using Variadic Args to a Simple function.
I also included a Simple function that uses plain old primitive args for comparison purposes.

## Results

```
============================================================================
velox/expression/benchmarks/VariadicBenchmark.cpprelative  time/iter  iters/s
============================================================================
simple                                                      79.76ns   12.54M
vectorVariadic                                              89.15ns   11.22M
simpleVariadic                                              88.65ns   11.28M
vectorVariadicLotsOfArgs                                     1.86us  536.79K
simpleVariadicLotsOfArgs                                     2.02us  495.56K
============================================================================
```

Not surprisingly the simple function without variadic args performs the best.

The simple function with variadic args still outperforms the vector function with a small number of
arguments (5), but shows a bit of a regression (~5%) with a large number of arguments (100).

I think this is due to the overhead of wrapping values in VectorOptionalValueAccessor in
VariadicView.  As an experiment I tried hard coding VariadicView to not do that wrapping (setting
returnsOptionalValues to false) and the simple function started outperforming the vector function
with a large number of arguments (it didn't significantly affect performance with a small number of
arguments).

Technically we only need to wrap the arguments in VectorOptionalValueAccessor in "callNullable",
but given the complexity for the user of having different argument types for "call" and "callNullable"
just for Variadic args, I'm not inclined to make that change unless it becomes a noticeable
performance bottleneck.

Reviewed By: laithsakka

Differential Revision: D34058182

fbshipit-source-id: 3ded324545b02f2c9fec670026907829d20c065e