Nested TRYs catch the same errors over and over (#1093)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1093

TRY expressions don't reset the errors in the EvalCtx, so if a nested TRY "catches" an error
and NULLs out the row, the upstream TRY will catch the same error again and NULL it out
again (even if there was a coalesce or some other expression that converts the NULL to
some non-NULL value).  In addition, with nested TRYs, if an expression in an earlier branch
threw an exception, the inner TRY will see and handle (null out that branch for) any rows
that saw errors in that earlier branch.

To handle this I've modified TryExpr to store the current state of errors in the EvalCtx when
it's invoked and reset the errors to null. When the TryExpr returns it resets the state of
errors back to what it was originally.  This way the parent TRY never sees errors the child
TRY already handled, and the child TRY doesn't see any errors the parent TRY is supposed
to handle.

Reviewed By: laithsakka

Differential Revision: D34403320

fbshipit-source-id: 839f20b9e832fc13d0400b56a596fa136230997c