Try expression does not NULL out constant vectors (#1098)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1098

If the child of a TRY expression produces a ConstantVector (e.g. because all of its inputs
are constant) TRY will call setNull on that vector which has no effect on a ConstantVector.

I added a special case, that returns a null ConstantVector when this happens.

I didn't handle the case where the result is a ConstantVector, but only some rows throw
exceptions as I don't think that's possible.  I added a check just in case (it's possible to
handle, I didn't want to needlessly complicate the code).

Reviewed By: laithsakka

Differential Revision: D34488476

fbshipit-source-id: 57f580600cc0d0258bd98e69f0ed98e4b4754dba