Simplify SIMD Comparison
Removes the concept of compareResult', an intermediate formm between a
SIMD vector from a comparison and the corresponding bit mask. Removes
the use of pext + movemask_epi8 and uses the direct movemask_pd and
movemask_ps instead. There is no movemask_epi16, so for a 16x16 vector
we still use movemask_epi8 + pext but this is not frequent.