Fix presto_cpp e2e test TestHiveAggregationQueries.testApproxDistinct (#1117)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1117

Casting to HyperLogLog failed after introducing a new logic in CastExpr that calls custom cast operators when the from-type or to-type is a custom type. HyperLogLog is a custom type but does not implement its own cast operator. Instead, it reused the cast logic for Varbinary. This diff makes HyperLogLogTypeFactories::getCastOperator() return a nullptr so that CastExpr falls through to use the cast logic of Varbinary as it did before.

Reviewed By: mbasmanova

Differential Revision: D34594007

fbshipit-source-id: 929ec8d19dddfc38ba67950abb1e30d46b7340c9