Resolve simple functions based on generality. (#1042)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1042

This diff completes the signature ranking work, and adds the e2e support
to resolving simple functions based on their priority which is determined using
the generality of the signature.

For example a function add(...) can be implemented as the following:

```
template<typename T>
struct AddFunc {

  bool call(int & out, int a , int b){
    out = a+b;
  }

  bool call(int & out, const arg_type<Variadic<int>>& inputs){
   out = 0;
   for(auto v: inputs){
      out+=*v;
   }
 }
};

registerFunction<AddFunc, int, int, int>({"add"});
registerFunction<AddFunc, int, Variadic<int>>{"add"};
```
In the case above we want to make sure that the first add is
picked when the inputs are just two integers. since both registrations would work.

Priority is assigned as the following:
Lower priorities are picked first during function resolutions. The computation works as the
following, each signature get a rank out of 4. Those ranks form initial Lattice
ordering.

rank 1: generic free and variadic free.
     e.g: int, int, int -> int.

rank 2: has variadic but generic free.
     e.g: Variadic<int> -> int.

rank 3: has generic but no variadic of generic.
     e.g: Generic<>, Generic<>,Generic<>  -> int.

rank 4: has variadic of generic.
     e.g: Variadic<Generic<>> -> int.

If two functions have the same rank, then concreteCount is used to
to resolve the ordering which is the number of concrete types
(other than variadic and generic) in the signature.

e.g: consider the two functions:
     1. int, Generic<>, Variadic<int> -> has rank 3. concreteCount =2
     2. int, Generic<>, Generic<>    -> has rank 3. concreteCount =1

in this case (2) is picked.
  e.g: (Generic<>, int) will be picked before (Generic<>, Generic<>)
  e.g: Variadic<Array<Generic<>>> is picked before Variadic<Generic<>>.

Differential Revision: D34233219

fbshipit-source-id: 197b8e760dc1e65024a0c2de3cf313a63a3c48f2