Allow users to provide `void call(...)` (#1129)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1129

The main idea is to allow users to provide a version of the
call|callNullable|callNullFree|callAscii methods returning void, and letting
the wrapper class detect and return true instead. This is based on the fact
that the absolute majority of simple functions has a:
```
    ....
    return true;
  }
```
at the end. Changes are encapsulated in the SimpleFunctionMetadata
class; it's api hasn't changed.

Performance-wise this translates to the exact same binary, but it allow us to
use these flags and provide a faster path in subsequent diffs in case we're
executing a function that never returns nulls.

Reviewed By: mbasmanova

Differential Revision: D34663928

fbshipit-source-id: b112ceb1fa84a394f8e346ba36b3d3c602ae4d6b