Add IOBufOutputStream
Adds an OutputStream that allows retrieving the content as
folly::IOBufs. This is for use in partitioned output to get zero copy
message passing between Velox and the exchange server in presto_cpp.

Changes the Position type of ByteStream to a number so as to be
compatible with C++ streams.