Fixes for timestamp in UnsafeRow{Dynamic,}Serializer
Summary:
Avoid duplicated logic for serialized fixed-width values by funneling
everything through serializedFixedWidth. As part of trying to clean this up, I
also removed a bunch of function implementations that only emitted errors. This
revealed that the dynamic interface was being called with things other than
Vectors, so I locked down the interface and updated callers in the tests to
call the static serializer instead in these cases.

Differential Revision: D34751687

fbshipit-source-id: 48f835fa01c3259f6726343a74605cc0448caf8b