Simplified UnsafeRow deserializer
Summary:
In an attempt to simplify the unsafe row desrializer, I started from the
simpler interface I suggested in my review of D34440865 and filled out the
implementation for all types. I haven't benchmarked this version  yet.

I made the existing deserializer tests templated so they could be used to test
multiple deserializer implementations and added a round-trip fuzzer that
generates a random complex type, generates a vector for that type, serializes
the vector, deserializes it, and confirms that the result is the same as the
original. This revealed some bugs related to timestamps and structs with zero
fields (see other diffs).

Differential Revision: D34751686

fbshipit-source-id: 065c341570ae56c538510667c7052fd7279eafba