Fixes for timestamp in UnsafeRow{Dynamic,}Serializer
Summary:
Avoid duplicated logic for serialized fixed-width values by funneling
everything through serializedFixedWidth. As part of trying to clean this up, I
also removed a bunch of function implementations that only emitted errors. This
revealed that the dynamic interface was being called with things other than
Vectors, so I locked down the interface and updated callers in the tests to
call the static serializer instead in these cases.

Differential Revision: D34751687

fbshipit-source-id: ccd5b0d1a5283e5a0255824587b1dd77c85db037