Simplified UnsafeRow deserializer (#1153)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1153

In an attempt to simplify the unsafe row deserializer, I started from the
simpler interface I suggested in my review of D34440865 and filled out the
implementation for all types. I haven't benchmarked this version  yet.

I made the existing deserializer tests templated so they could be used to test
multiple deserializer implementations and added a round-trip fuzzer that
generates a random complex type, generates a vector for that type, serializes
the vector, deserializes it, and confirms that the result is the same as the
original. This revealed some bugs related to timestamps and structs with zero
fields (see other diffs).

I've called this "UnsafeRow24Deserializer" to clarify that it implements UnsafeRow as expected by Spark 2.4.

Reviewed By: miaoever

Differential Revision: D34751686

fbshipit-source-id: d179aec0e828471bcc76ae6ed75c5e4274b06b3d