Fixes for timestamp in UnsafeRow{Dynamic,}Serializer (#1158)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1158

Avoid duplicated logic for serialized fixed-width values by funneling
everything through serializedFixedWidth. As part of trying to clean this up, I
also removed a bunch of function implementations that only emitted errors. This
revealed that the dynamic interface was being called with things other than
Vectors, so I locked down the interface and updated callers in the tests to
call the static serializer instead in these cases.

Reviewed By: miaoever

Differential Revision: D34751687

fbshipit-source-id: 2bd72638166298514caa46496f8f08d67ee2de27