Reduce usage of DriverCtx::execCtx (#1166)
Summary:
FilterProject operator uses ExecCtx from the Driver for expression evaluation.
This means that memory usage that happens during expression evaluation is
tracked in the Driver's memory pool and is not being attributed to the
operator. To fix this, we'll remove ExecCtx from the Driver and have each
operator create its own using operator's memory pool. This is a preparation
step to reduce usage of DriverCtx::execCtx.

Introduce DriverCtx::queryConfig and replace driverCtx->execCtx->queryCtx
()->config() calls with driverCtx->queryConfig().

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1166

Reviewed By: spershin

Differential Revision: D34770035

Pulled By: mbasmanova

fbshipit-source-id: 9ad03b5bfc41f729fd6bafc901d44bc91e3991f0