Optimize deserialization of arrays of fixed-width elements
Summary:
This diff adds fast-paths for arrays of fixed-width types (`DeserializeFixedWidthArrayElements`). `DeserializeArray` now has a type switch that dispatches to this function. `arrayElementStride` is kept for now because it is used by `DeserializeMap`, but the next diff rewrites `DeserializeMap` by re-using `DeserializeArray`.

The logic in `DeserializeFixedWidthArrayElements` to calculate the pointer to the next element would be drastically simpler if it could be written in coroutine style. Unfortunately that is not possible.

DeserializeArray is getting a little long at ~60 lines, but it's within reason.

Reviewed By: miaoever

Differential Revision: D34804831

fbshipit-source-id: 16dee792ebb688cfef23c2ccb48d171158c4f57b