Move execCtx from Driver to Operator (#1174)
Summary:
FilterProject operator uses ExecCtx from the Driver for expression evaluation.
This means that memory usage that happens during expression evaluation is
tracked in the Driver's memory pool and is not being attributed to the
operator. To fix this, we move ExecCtx to the Operator and use operator's
memory pool to create it.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1174

Reviewed By: pedroerp

Differential Revision: D34789360

Pulled By: mbasmanova

fbshipit-source-id: 8faeaabe65a15a1f254925dde3e5f4fd1cc33a0f