Add serialized validation function for dense HLL (#1186)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1186

Dense Hll deserializing and cardinality function assume that the format of the input is correct and do not check for possibly wrong values that can break the functions. These values cause the function to break while running leading to USBAN errors that cannot be caught.

These problems stem from the input being sometimes wrong–either too short or with values that are not in correct ranges.

Added optional function to check if a string is in valid Dense HLL format.

There don't seem to be problems like this with SparseHLL. They don't use values from the input for shifts like DenseHLL does and they don't seem to read it the same way. SparseHll is also a much simpler format, just composed of a bucket to the count. It doesn't appear to need any extra checks. It's possible that a random varbinary is not meant to be SparseHll but is a valid SparseHll however that is not something we can prevent.

Differential Revision: D34651255

fbshipit-source-id: 0ac676f66ef9c647b5ddea21591a751d8294933c