Add IOBufOutputStream (#1123)
Summary:
Adds an OutputStream that allows retrieving the content as
folly::IOBufs. This is for use in partitioned output to get zero copy
message passing between Velox and the exchange server in presto_cpp.

Changes the Position type of ByteStream to a number so as to be
compatible with C++ streams.

Pull Request resolved: https://github.com/facebookincubator/velox/pull/1123

Reviewed By: mbasmanova

Differential Revision: D34628220

Pulled By: oerling

fbshipit-source-id: f638ffda6ceb86e8df21063b0f43e70bdb0be8eb