Support random_access_iterator requirements in MapView::Iterator (#1221)
Summary:
Pull Request resolved: https://github.com/facebookincubator/velox/pull/1221

As the title suggests, this diff makes it so that MapView::Iterator meets the requirements
of a random_access_iterator.  This lets std algorithms on instances of the iterator like
distance know that they can use constant time implementations rather than linear time.

The new functions include:
In the IndexBasedIterator:
pre-decrement
post-decrement
+=
-=
In the MapView::Iterator
iterator + int
int + iterator
iterator - int
iterator - iterator
[n] to access the value at iterator + n

In addition I added calls to the validateBounds() function in IndexBasedIterator when
dereferencing the iterator to ensure we don't accidentally dereference values off the end
of the current collection.

I also added a static check to the Iterator to ensure it's properly declaring its
iterator_category (Iterators don't inherit this from the parent, which it seems like the code
was assuming before).

Differential Revision: D34911295

fbshipit-source-id: 9f1f7cfdac9e929fac3469c3c3ed11255a83c6fa