Support random_access_iterator requirements in VariadicView::Iterator
Summary:
As the title suggests, this diff makes it so that VariadicView::Iterator meets the
requirements of a random_access_iterator.  This lets std algorithms on instances of the
iterator like distance know that they can use constant time implementations rather than
linear time.

The new functions include:
In the IndexBasedIterator:
pre-decrement
post-decrement
+=
-=
In the VariadicView::Iterator
iterator + int
int + iterator
iterator - int
iterator - iterator
[n] to access the value at iterator + n

In addition I added calls to the validateBounds() function in IndexBasedIterator when
dereferencing the iterator to ensure we don't accidentally dereference values off the end
of the current collection.

I also added a static check to the Iterator to ensure it's properly declaring its
iterator_category (Iterators don't inherit this from the parent, which it seems like the code
was assuming before).

Lastly, I removed the const from the offset_ field in the Iterator.  While running the tests I
stumbled on the fact the default assignment operator was deleted because this field is
const.  This was not intentional so I removed the restriction (though it's effectively const).

Differential Revision: D34913086

fbshipit-source-id: 6a89ae5351196c6267d38eecacf4d238c2485810