Remove extra copying in PartitionedOutput/Exchange
Writes the result of serialization into SerializedPage in
PartitionedOutput and makes SerializedPage the content type of
PartitionedOutputBufferManager. Uses folly::IOBuf as
the content type of SerializedPage, so that IOBufs received from
Proxygen in presto_cpp can be owned by the receiving ExchangeQueue without copy.

Since the serialization is finalized in PartitionedOutput, its CPU
time and transient memory are accounted for the Task, which is not
always the case if this is done in the presto_cpp TaskManager. The
serialization is done in MappedMemory and the ownership is handed to
the resulting IOBufs without copy.