Fix LazyVector in MergeJoin
Loads LazyVectors in CallbackSink. The load is scoped to the rows that
survive all filters. In general, this node passes data outside of the
pipeline, which means that after this point there is no guarantee for
the lifetime of the readers of the LazyVectors.

Loads all lazy vectors of inputs buffered by MergeJoin. The buffering
takes place when a left batch ends with a hit.

Adds a utility to load all LazyVectors reachable from a
RowVector. Note that even when a vector does not contain unloaded
LazyVectors, loadedVector() must be called to initialize wrappers
around LazyVectors that may have been loaded via some other path.
Separates out the logic for loading a selected set of top level rows
from LazyVectors that may be arbitrarily wrapped (ensureloadedRows).